Remove `git2` integration
