+++ packages/ploys/Cargo.toml
@@ -10,7 +10,7 @@ edition = "2021"
 
 [features]
 default = ["git", "github"]
-git = ["dep:gix", "dep:git2"]
+git = ["dep:gix"]
 github = ["dep:ureq"]
 
 [dependencies]
@@ -23,11 +23,6 @@ time = { version = "0.3.36", features = ["serde", "formatting", "parsing"] }
 toml_edit = { version = "0.22.14", features = ["serde"] }
 ureq = { version = "2.7.1", features = ["json"], optional = true }
 url = "2.4.0"
-
-[dependencies.git2]
-version = "0.19.0"
-features = ["vendored-libgit2", "vendored-openssl"]
-optional = true
 
 [dev-dependencies]
 indoc = "2.0.5"
