+++ packages/ploys/src/project/source/git/mod.rs
@@ -3,26 +3,24 @@
 //! This module contains the utilities related to local Git project management.
 
 mod error;
-mod git2;
-mod gix;
 
+use std::io;
 use std::path::{Path, PathBuf};
 
+use gix::remote::Direction;
+use gix::traverse::tree::Recorder;
+use gix::Repository;
 use url::Url;
 
-pub use self::error::{Error, GixError};
-pub use self::git2::Git2;
-pub use self::gix::Gix;
+pub use self::error::Error;
 
 use super::revision::Revision;
 use super::Source;
 
 /// The local Git repository source.
-pub enum Git {
-    /// The `gix` source for basic `git` operations.
-    Gix(Gix),
-    /// The `git2` source for advanced `git` operations.
-    Git2(Git2),
+pub struct Git {
+    repository: Repository,
+    revision: Revision,
 }
 
 impl Git {
@@ -31,25 +29,22 @@ impl Git {
     where
         P: AsRef<Path>,
     {
-        Ok(Self::Gix(Gix::new(path)?))
+        Ok(Self {
+            repository: gix::open(path.as_ref())?,
+            revision: Revision::Head,
+        })
     }
 }
 
 impl Git {
     /// Gets the revision.
     pub fn revision(&self) -> &Revision {
-        match self {
-            Self::Gix(gix) => gix.revision(),
-            Self::Git2(git2) => git2.revision(),
-        }
+        &self.revision
     }
 
     /// Sets the revision.
     pub fn set_revision(&mut self, revision: impl Into<Revision>) {
-        match self {
-            Self::Gix(gix) => gix.set_revision(revision),
-            Self::Git2(git2) => git2.set_revision(revision),
-        }
+        self.revision = revision.into();
     }
 
     /// Builds the source with the given revision.
@@ -71,33 +66,79 @@ impl Source for Git {
     }
 
     fn get_name(&self) -> Result<String, Self::Error> {
-        match self {
-            Self::Gix(gix) => gix.get_name(),
-            Self::Git2(git2) => git2.get_name(),
+        let path = self.repository.path().join("..").canonicalize()?;
+
+        if let Some(file_stem) = path.file_stem() {
+            return Ok(file_stem.to_string_lossy().to_string());
         }
+
+        Err(Error::Io(io::Error::new(
+            io::ErrorKind::Other,
+            "Invalid directory",
+        )))
     }
 
     fn get_url(&self) -> Result<Url, Self::Error> {
-        match self {
-            Self::Gix(gix) => gix.get_url(),
-            Self::Git2(git2) => git2.get_url(),
+        match self
+            .repository
+            .find_default_remote(Direction::Push)
+            .transpose()?
+        {
+            Some(remote) => match remote.url(Direction::Push) {
+                Some(url) => Ok(url
+                    .to_bstring()
+                    .to_string()
+                    .parse()
+                    .expect("A repository URL should be valid")),
+                None => Err(Error::remote_not_found()),
+            },
+            None => Err(Error::remote_not_found()),
         }
     }
 
     fn get_files(&self) -> Result<Vec<PathBuf>, Self::Error> {
-        match self {
-            Self::Gix(gix) => gix.get_files(),
-            Self::Git2(git2) => git2.get_files(),
-        }
+        let spec = self.revision.to_string();
+        let tree = self
+            .repository
+            .rev_parse_single(&*spec)?
+            .object()?
+            .peel_to_tree()?;
+
+        let mut recorder = Recorder::default();
+
+        tree.traverse().breadthfirst(&mut recorder)?;
+
+        let mut entries = recorder
+            .records
+            .into_iter()
+            .filter(|entry| entry.mode.is_blob())
+            .map(|entry| PathBuf::from(entry.filepath.to_string()))
+            .collect::<Vec<_>>();
+
+        entries.sort();
+
+        Ok(entries)
     }
 
     fn get_file_contents<P>(&self, path: P) -> Result<Vec<u8>, Self::Error>
     where
         P: AsRef<Path>,
     {
-        match self {
-            Self::Gix(gix) => gix.get_file_contents(path),
-            Self::Git2(git2) => git2.get_file_contents(path),
+        let spec = self.revision.to_string();
+        let mut tree = self
+            .repository
+            .rev_parse_single(&*spec)?
+            .object()?
+            .peel_to_tree()?;
+
+        let entry = tree
+            .peel_to_entry_by_path(path)?
+            .ok_or_else(|| io::Error::from(io::ErrorKind::NotFound))?;
+
+        if entry.mode().is_blob() {
+            Ok(entry.object()?.detached().data)
+        } else {
+            Err(io::Error::from(io::ErrorKind::NotFound))?
         }
     }
 }
